Labfiles/01-build-agent-portal-and-vscode/Python/requirements.txt
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
